+++ src/repositories/contracts/FoodItemRepository.ts
@@ -1,0 +1,7 @@
+import { FoodItem } from "@/domain";
+import { Repository } from "@/repositories/Repository";
+
+export interface FoodItemRepository extends Repository<FoodItem, string> {
+  findByName(query: string): FoodItem[];
+}
+